#33012: Complete overhaul of http-client and install-tics
Now using `@tiobe/http-client` and `@tiobe/install-tics` to retrieve data from the viewer and to get the install TICS url from the viewer. Rewrote the testing for it and fixed the proxy setup, which now passes the local tests as well as the unit tests. To do this I had to revert some libraries to get it to work.